@@ -1,19 +1,22 @@
 use bytes::Bytes;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
+use std::future;
+use std::pin::Pin;
 use std::sync::mpsc::{Receiver, TryRecvError};
 use std::sync::{Arc, Mutex};
+use std::task::{Context, Poll};
 use std::time::Duration;
+use tokio::sync::watch;
 use tokio::time::Instant;
 
+use crate::async_impl::body::ResponseBody;
 use crate::error::{BoxError, Error, Kind};
 use crate::Body;
 use bytes::Buf;
-use futures_util::future;
 use h3::client::SendRequest;
 use h3_quinn::{Connection, OpenStreams};
 use http::uri::{Authority, Scheme};
 use http::{Request, Response, Uri};
-use hyper::Body as HyperBody;
 use log::trace;
 
 pub(super) type Key = (Scheme, Authority);
@@ -23,23 +26,89 @@ pub struct Pool {
     inner: Arc<Mutex<PoolInner>>,
 }
 
+struct ConnectingLockInner {
+    key: Key,
+    pool: Arc<Mutex<PoolInner>>,
+}
+
+/// A lock that ensures only one HTTP/3 connection is established per host at a
+/// time. The lock is automatically released when dropped.
+pub struct ConnectingLock(Option<ConnectingLockInner>);
+
+/// A waiter that allows subscribers to receive updates when a new connection is
+/// established or when the connection attempt fails. For example, whe
+/// connection lock is dropped due to an error.
+pub struct ConnectingWaiter {
+    receiver: watch::Receiver<Option<PoolClient>>,
+}
+
+pub enum Connecting {
+    /// A connection attempt is already in progress.
+    /// You must subscribe to updates instead of initiating a new connection.
+    InProgress(ConnectingWaiter),
+    /// The connection lock has been acquired, allowing you to initiate a
+    /// new connection.
+    Acquired(ConnectingLock),
+}
+
+impl ConnectingLock {
+    fn new(key: Key, pool: Arc<Mutex<PoolInner>>) -> Self {
+        Self(Some(ConnectingLockInner { key, pool }))
+    }
+
+    /// Forget the lock and return corresponding Key
+    fn forget(mut self) -> Key {
+        // Unwrap is safe because the Option can be None only after dropping the
+        // lock
+        self.0.take().unwrap().key
+    }
+}
+
+impl Drop for ConnectingLock {
+    fn drop(&mut self) {
+        if let Some(ConnectingLockInner { key, pool }) = self.0.take() {
+            let mut pool = pool.lock().unwrap();
+            pool.connecting.remove(&key);
+            trace!("HTTP/3 connecting lock for {:?} is dropped", key);
+        }
+    }
+}
+
+impl ConnectingWaiter {
+    pub async fn receive(mut self) -> Option<PoolClient> {
+        match self.receiver.wait_for(Option::is_some).await {
+            // unwrap because we already checked that option is Some
+            Ok(ok) => Some(ok.as_ref().unwrap().to_owned()),
+            Err(_) => None,
+        }
+    }
+}
+
 impl Pool {
     pub fn new(timeout: Option<Duration>) -> Self {
         Self {
             inner: Arc::new(Mutex::new(PoolInner {
-                connecting: HashSet::new(),
+                connecting: HashMap::new(),
                 idle_conns: HashMap::new(),
                 timeout,
             })),
         }
     }
 
-    pub fn connecting(&self, key: Key) -> Result<(), BoxError> {
+    /// Aqcuire a connecting lock. This is to ensure that we have only one HTTP3
+    /// connection per host.
+    pub fn connecting(&self, key: &Key) -> Connecting {
         let mut inner = self.inner.lock().unwrap();
-        if !inner.connecting.insert(key.clone()) {
-            return Err(format!("HTTP/3 connecting already in progress for {:?}", key).into());
+
+        if let Some(sender) = inner.connecting.get(key) {
+            Connecting::InProgress(ConnectingWaiter {
+                receiver: sender.subscribe(),
+            })
+        } else {
+            let (tx, _) = watch::channel(None);
+            inner.connecting.insert(key.clone(), tx);
+            Connecting::Acquired(ConnectingLock::new(key.clone(), Arc::clone(&self.inner)))
         }
-        return Ok(());
     }
 
     pub fn try_pool(&self, key: &Key) -> Option<PoolClient> {
@@ -70,34 +139,47 @@ impl Pool {
 
     pub fn new_connection(
         &mut self,
-        key: Key,
+        lock: ConnectingLock,
         mut driver: h3::client::Connection<Connection, Bytes>,
         tx: SendRequest<OpenStreams, Bytes>,
     ) -> PoolClient {
         let (close_tx, close_rx) = std::sync::mpsc::channel();
         tokio::spawn(async move {
             if let Err(e) = future::poll_fn(|cx| driver.poll_close(cx)).await {
-                trace!("poll_close returned error {:?}", e);
+                trace!("poll_close returned error {e:?}");
                 close_tx.send(e).ok();
             }
         });
 
         let mut inner = self.inner.lock().unwrap();
 
-        let client = PoolClient::new(tx);
-        let conn = PoolConnection::new(client.clone(), close_rx);
-        inner.insert(key.clone(), conn);
-
         // We clean up "connecting" here so we don't have to acquire the lock again.
-        let existed = inner.connecting.remove(&key);
-        debug_assert!(existed, "key not in connecting set");
+        let key = lock.forget();
+        let Some(notifier) = inner.connecting.remove(&key) else {
+            unreachable!("there should be one connecting lock at a time");
+        };
+        let client = PoolClient::new(tx);
+
+        // Send the client to all our awaiters
+        let pool_client = if let Err(watch::error::SendError(Some(unsent_client))) =
+            notifier.send(Some(client.clone()))
+        {
+            // If there are no awaiters, the client is returned to us. As a
+            // micro optimisation, let's reuse it and avoid clonning.
+            unsent_client
+        } else {
+            client.clone()
+        };
+
+        let conn = PoolConnection::new(pool_client, close_rx);
+        inner.insert(key, conn);
 
         client
     }
 }
 
 struct PoolInner {
-    connecting: HashSet<Key>,
+    connecting: HashMap<Key, watch::Sender<Option<PoolClient>>>,
     idle_conns: HashMap<Key, PoolConnection>,
     timeout: Option<Duration>,
 }
@@ -105,7 +187,7 @@ struct PoolInner {
 impl PoolInner {
     fn insert(&mut self, key: Key, conn: PoolConnection) {
         if self.idle_conns.contains_key(&key) {
-            trace!("connection already exists for key {:?}", key);
+            trace!("connection already exists for key {key:?}");
         }
 
         self.idle_conns.insert(key, conn);
@@ -125,9 +207,19 @@ impl PoolClient {
     pub async fn send_request(
         &mut self,
         req: Request<Body>,
-    ) -> Result<Response<HyperBody>, BoxError> {
+    ) -> Result<Response<ResponseBody>, BoxError> {
+        use hyper::body::Body as _;
+
         let (head, req_body) = req.into_parts();
-        let req = Request::from_parts(head, ());
+        let mut req = Request::from_parts(head, ());
+
+        if let Some(n) = req_body.size_hint().exact() {
+            if n > 0 {
+                req.headers_mut()
+                    .insert(http::header::CONTENT_LENGTH, n.into());
+            }
+        }
+
         let mut stream = self.inner.send_request(req).await?;
 
         match req_body.as_bytes() {
@@ -141,12 +233,9 @@ impl PoolClient {
 
         let resp = stream.recv_response().await?;
 
-        let mut resp_body = Vec::new();
-        while let Some(chunk) = stream.recv_data().await? {
-            resp_body.extend(chunk.chunk())
-        }
+        let resp_body = crate::async_impl::body::boxed(Incoming::new(stream, resp.headers()));
 
-        Ok(resp.map(|_| HyperBody::from(resp_body)))
+        Ok(resp.map(|_| resp_body))
     }
 }
 
@@ -176,6 +265,52 @@ impl PoolConnection {
             Err(TryRecvError::Empty) => false,
             Err(TryRecvError::Disconnected) => true,
             Ok(_) => true,
+        }
+    }
+}
+
+struct Incoming<S, B> {
+    inner: h3::client::RequestStream<S, B>,
+    content_length: Option<u64>,
+}
+
+impl<S, B> Incoming<S, B> {
+    fn new(stream: h3::client::RequestStream<S, B>, headers: &http::header::HeaderMap) -> Self {
+        Self {
+            inner: stream,
+            content_length: headers
+                .get(http::header::CONTENT_LENGTH)
+                .and_then(|h| h.to_str().ok())
+                .and_then(|v| v.parse().ok()),
+        }
+    }
+}
+
+impl<S, B> http_body::Body for Incoming<S, B>
+where
+    S: h3::quic::RecvStream,
+{
+    type Data = Bytes;
+    type Error = crate::error::Error;
+
+    fn poll_frame(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context,
+    ) -> Poll<Option<Result<hyper::body::Frame<Self::Data>, Self::Error>>> {
+        match futures_core::ready!(self.inner.poll_recv_data(cx)) {
+            Ok(Some(mut b)) => Poll::Ready(Some(Ok(hyper::body::Frame::data(
+                b.copy_to_bytes(b.remaining()),
+            )))),
+            Ok(None) => Poll::Ready(None),
+            Err(e) => Poll::Ready(Some(Err(crate::error::body(e)))),
+        }
+    }
+
+    fn size_hint(&self) -> hyper::body::SizeHint {
+        if let Some(content_length) = self.content_length {
+            hyper::body::SizeHint::with_exact(content_length)
+        } else {
+            hyper::body::SizeHint::default()
         }
     }
 }
