@@ -62,7 +62,7 @@ impl Response {
     ///
     /// - The server didn't send a `content-length` header.
     /// - The response is compressed and automatically decoded (thus changing
-    ///  the actual decoded length).
+    ///   the actual decoded length).
     pub fn content_length(&self) -> Option<u64> {
         self.headers()
             .get(http::header::CONTENT_LENGTH)?
